add ft_stcmp

+++ timer.c
@@ -6,15 +6,15 @@
 /*   By: zbakkas <zouhirbakkas@gmail.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/04 17:19:51 by zbakkas           #+#    #+#             */
-/*   Updated: 2024/06/04 17:35:07 by zbakkas          ###   ########.fr       */
+/*   Updated: 2024/06/05 10:59:03 by zbakkas          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-long int	timer(t_args *args)
+size_t	timer(t_args *args)
 {
-	long	res;
+	size_t	res;
 
 	res = get_current_time() - args->start;
 	return (res);
@@ -28,7 +28,7 @@ size_t	get_current_time(void)
 	return (time.tv_sec * 1000 + time.tv_usec / 1000);
 }
 
-int	ft_usleep(int milliseconds, t_philo *philo)
+int	ft_usleep(size_t milliseconds, t_philo *philo)
 {
 	size_t	start;
